requirements.txt
selenium==4.0.0
webdriver-manager==4.0.1
numpy
mysql-connector-python
matplotlib
python-dotenv
python-socketio